Move types to their own file
